Mobileapp/requirements.txt
Flask
pandas
openpyxl
python-docx
scikit-learn
requests
